elf128 : fix 128-bit specific code for XLEN independant compilation

+++ binutils/readelf.c
@@ -1585,7 +1585,7 @@ dump_relocations (Filedata *          filedata,
 	      rel_size);
       for (i = 0; i < rel_size; i++) {
         if (!is_32bit_elf && !is_64bit_elf)
-          printf (format, (unsigned long) (relrs[i] >> 64));
+          printf (format, (unsigned long) ((__uint128_t) relrs[i] >> 64));
         printf (format, (unsigned long) relrs[i]);
       }
       free (relrs);
@@ -1676,8 +1676,8 @@ dump_relocations (Filedata *          filedata,
 	  printf (do_wide
 		  ? "%16.16" BFD_VMA_FMT "x%16.16" BFD_VMA_FMT "x  %16.16" BFD_VMA_FMT "x%16.16" BFD_VMA_FMT "x "
 		  : "%12.12" BFD_VMA_FMT "x%12.12" BFD_VMA_FMT "x  %12.12" BFD_VMA_FMT "x%12.12" BFD_VMA_FMT "x ",
-      (unsigned long) (offset >> 64), (unsigned long) offset,
-      (unsigned long) (inf >> 64), (unsigned long) inf);
+      (unsigned long) ((__uint128_t) offset >> 64), (unsigned long) offset,
+      (unsigned long) ((__uint128_t) inf >> 64), (unsigned long) inf);
 	}
 
       switch (filedata->file_header.e_machine)
